`Simple' change to eliminate dy_cmdchn, dy_cmdecho globals. Along the way fixed a bunch of warnings about converting 64-bit addresses to 32-bit ints, fixed an error in va_start/va_end usage in dylib_io.c and dylib_errs.c, and made the examples work again.

+++ DyLP/src/OsiDylp/OsiDylpWarmStartBasis.cpp
@@ -76,7 +76,7 @@ using std::vector ;
 */
 
 #define STATPERBYTE 4
-#define STATALLOCUNIT sizeof(int)
+#define STATALLOCUNIT (static_cast<int>(sizeof(int)))
 #define STATBYTES(zz_ns_zz) \
   (((zz_ns_zz+STATALLOCUNIT*STATPERBYTE-1)/(STATALLOCUNIT*STATPERBYTE))* \
    STATALLOCUNIT)
@@ -511,8 +511,8 @@ void ODWSB::deleteRows (int rawTgtCnt, const int *rawTgts)
   int *endUnique ;
   std::sort(first,last) ;
   endUnique = std::unique(first,last) ;
-  int tgtCnt = endUnique-first ;
-  compressRows(tgtCnt,tgts) ;
+  ptrdiff_t tgtCnt = endUnique-first ;
+  compressRows(static_cast<int>(tgtCnt),tgts) ;
   delete [] tgts ;
 
 # if ODSI_PARANOIA >= 2
